them danh sach dang ki ho nhan vien

+++ lib/app/manager_leave_form/model/day_off_letter_manager_model.dart
@@ -1,7 +1,7 @@
 class DayOffLettersManagerModel {
   int? empID;
   int? regID;
-  int? period;
+  double? period;
   String? startDate;
   String? regDate;
   String? type;
